Refactor: Create shared_models and remove duplicate code
