--- src/test/config.py
@@ -1,19 +1,0 @@
-# Path to the image used to test allowedFile with small resolution
-API_IMAGE1 = "allowedFile_test1.png"
-# Path to the image used to test allowedFile with too large file
-API_IMAGE2 = "allowedFile_test2.png"
-# Path to the image used to test allowedFile with wrong format
-API_IMAGE3 = "allowedFile_test3.jpg"
-# Path to the image used to test allowedFile with correct input
-API_IMAGE4 = "allowedFile_test4.png"
-# Path to the image used to test the white_image function
-API_IMAGE5 = "white_image.png"
-# test file for customvision module
-CV_TEST_IMAGE = "cv_testfile.png"
-# Contains in-order sequence to the directory with images used to test
-# the allowedFile function
-API_PATH_DATA = ["..", "data"]
-# Name of the labels for the DB tests
-LABELS = "label1, label2, label3"
-# How long the test games last
-STATE = "playing"
