+++ src/utilities/languages.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Language(str, Enum):
+    Norwegian = "NO"
+    English = "EN"
